+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features / enhancements
 
 * [cli] add maintenance option to command-line admin tool
+* [core] add support for custom plans
 
 ### Fixes
 
@@ -79,6 +80,7 @@
 * [core] fix role checks on user unjoin (test was failing and email not sent)
 * [helpdesk] handle some error cases
 * [booking] show old client account used for a booking, in booking edition in case user switched to new client
+>>>>>>> develop
 
 ## 2.7.0